feat: SearchVC에 로딩 인디케이터 추가
